Merged with develop

+++ Common/src/matrix_structure.cpp
@@ -1192,7 +1192,7 @@ unsigned long CSysMatrix::Jacobi_Smoother(const CSysVector & b, CSysVector & x, 
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned long) i;
   
 }
 
@@ -1496,7 +1496,7 @@ unsigned long CSysMatrix::ILU0_Smoother(const CSysVector & b, CSysVector & x, CM
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned int) i;
   
 }
 
@@ -1659,7 +1659,7 @@ unsigned long CSysMatrix::LU_SGS_Smoother(const CSysVector & b, CSysVector & x, 
     cout << "# Iteration = " << i << ": |res|/|res0| = "  << norm_r/norm0 << ".\n" << endl;
   }
   
-  return i;
+  return (unsigned int) i;
   
 }
